requirements.txt
discord.py
google-genai
supabase
python-dotenv
requests
flask